Fix: Create MainDashboardFrame and replace where needed
