+++ components/dashboards/DashboardMenu.jsx
@@ -20,7 +20,7 @@ const DashboardMenu = (props) => {
       <div className='flex flex-col bg-primary-brownPod700 gap-7 p-7 mt-auto mb-auto h-[80%] text-primary-brownPod500 font-semibold z-20 text-xl'>
         <Link href='/profesional/perfil'>MI PERFIL</Link>
         <Link href='/profesional/asesores'>MIS ASESORES</Link>
-        <Link href='/'>MI EXPEDIENTE</Link>
+        <Link href='/profesional/expediente'>MI EXPEDIENTE</Link>
         <Link href='/'>AGENDA</Link>
         <Link href='/'>METODOS DE PAGO</Link>
         <Link href='/' className='text-right' onClick={signOut}>
