+++ pages/asesor/index.jsx
@@ -14,6 +14,8 @@ import SessionCard from '@/components/dashboards/SessionCard'
 import FormSquareButton from '@/components/forms/FormSquareButton'
 import SquareLink from '@/components/SquareLink'
 import ConsultantBanner from '@/components/dashboards/ConsultantBanner'
+import MainDashboardFrame from '@/components/dashboards/MainDashboardFrame'
+import DashboardNavbar from '@/components/dashboards/DashboardNavbar'
 
 import { jwtDecode } from 'jwt-decode'
 
@@ -45,18 +47,7 @@ const ConsultantDashboard = () => {
   }
 
   return (
-    <main className='box-content bg-contrast-slateGray50 min-h-screen w-full flex flex-col'>
-      {/* Navbar */}
-      <div className='bg-primary-brownPod800 w-full flex flex-row justify-center'>
-        <Image
-          className='p-3'
-          src='/assets/logos/logo-blanco.png'
-          alt='GuapoApp Logo'
-          width={110}
-          height={110}
-          priority
-        />
-      </div>
+    <MainDashboardFrame>
       {/* Principal */}
       <div className='flex flex-row w-full gap-10 justify-center z-10 px-10'>
         {/* Menu y sección izquierda */}
@@ -133,7 +124,7 @@ const ConsultantDashboard = () => {
       <div className='h-24 absolute -bottom-0 w-full z-0'>
         <CurvedBackground color='bg-primary-brownPod700' />
       </div>
-    </main>
+    </MainDashboardFrame>
   )
 }
 export default ConsultantDashboard
