+++ pages/profesional/asesores/index.jsx
@@ -14,7 +14,8 @@ import SessionCard from '@/components/dashboards/SessionCard'
 import FormSquareButton from '@/components/forms/FormSquareButton'
 import SquareLink from '@/components/SquareLink'
 import ConsultantCard from '@/components/dashboards/ConsultantCard'
-
+import MainDashboardFrame from '@/components/dashboards/MainDashboardFrame'
+import DashboardNavbar from '@/components/dashboards/DashboardNavbar'
 import DashboardMenu from '@/components/dashboards/DashboardMenu'
 
 import { useEffect, useState } from 'react'
@@ -48,18 +49,7 @@ const ProfessionalConsultants = () => {
   }
 
   return (
-    <main className='box-content bg-contrast-slateGray50 min-h-screen w-full flex flex-col'>
-      {/* Navbar */}
-      <div className='bg-primary-brownPod800 w-full flex flex-row justify-center'>
-        <Image
-          className='p-3'
-          src='/assets/logos/logo-blanco.png'
-          alt='GuapoApp Logo'
-          width={110}
-          height={110}
-          priority
-        />
-      </div>
+    <MainDashboardFrame>
       {/* Principal */}
       <section className='flex flex-row items-center justify-center px-10 z-10'>
         {/* Sección Izquierda || Menú */}
@@ -134,7 +124,7 @@ const ProfessionalConsultants = () => {
       <div className='h-24 absolute -bottom-0 w-full z-0'>
         <CurvedBackground color='bg-primary-brownPod700' />
       </div>
-    </main>
+    </MainDashboardFrame>
   )
 }
