+++ pages/profesional/perfil.jsx
@@ -13,6 +13,8 @@ import HourPicker from '@/components/dashboards/HourPicker'
 import SessionCard from '@/components/dashboards/SessionCard'
 import FormSquareButton from '@/components/forms/FormSquareButton'
 import SquareLink from '@/components/SquareLink'
+import MainDashboardFrame from '@/components/dashboards/MainDashboardFrame'
+import DashboardNavbar from '@/components/dashboards/DashboardNavbar'
 
 import { useForm } from 'react-hook-form'
 import { useRouter } from 'next/router'
@@ -59,18 +61,7 @@ const ProfessionalProfile = () => {
   }
 
   return (
-    <main className='box-content bg-contrast-slateGray50 min-h-screen w-full flex flex-col'>
-      {/* Navbar */}
-      <div className='bg-primary-brownPod800 w-full flex flex-row justify-center'>
-        <Image
-          className='p-3'
-          src='/assets/logos/logo-blanco.png'
-          alt='GuapoApp Logo'
-          width={110}
-          height={110}
-          priority
-        />
-      </div>
+    <MainDashboardFrame>
       {/* Principal */}
       <div className='flex flex-row w-full gap-14 justify-center z-10'>
         {/* Menu y sección izquierda */}
@@ -182,7 +173,7 @@ const ProfessionalProfile = () => {
       <div className='h-24 absolute -bottom-0 w-full z-0'>
         <CurvedBackground color='bg-primary-brownPod700' />
       </div>
-    </main>
+    </MainDashboardFrame>
   )
 }
